requirements.txt
python-dateutil>=2.8.2
pytz>=2022.1

SpeechRecognition>=3.8.1
PyAudio>=0.2.11
google-cloud-speech>=2.16.2
google-cloud-texttospeech>=2.12.0

google-api-python-client>=2.65.0
google-auth>=2.9.1
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=0.5.2

langchain>=0.1.0
langchain-openai>=0.0.2
openai>=1.3.0

pygame>=2.5.0

requests>=2.28.1
urllib3>=1.26.12
certifi>=2022.12.7
charset-normalizer>=2.1.1
idna>=3.4

dotenv